Drawer

+++ server/controllers/authController.js
@@ -78,11 +78,23 @@ exports.destroySession = (req, res, next) => {
 };
 
 exports.deleteUser = async (req, res) => {
+  const { userId } = req.body;
   try {
-    console.log(req.params.id);
-    await Comment.deleteMany({ author: req.params.id });
-    await Entry.deleteMany({ author: req.params.id });
-    await User.deleteMany({ _id: req.params.id });
+    if (req.session.user.id === req.params.id && req.session.user.role ===1) {
+      console.log(11111111, req.session.user)
+      await Comment.deleteMany({ author: req.params.id });
+      await Entry.deleteMany({ author: req.params.id });
+      await User.deleteMany({ _id: req.params.id });
+    }
+    if (req.session.user.role === 0 && req.params.id !== req.session.user.id) {
+      console.log(2222222222)
+
+      await Comment.deleteMany({ author: req.params.id });
+      await Entry.deleteMany({ author: req.params.id });
+      await User.deleteMany({ _id: req.params.id });
+    }
+    console.log(33333333)
+
   } catch (error) {
     console.log(error.message);
   }
@@ -105,16 +117,3 @@ exports.editUserProfilePicture = async (req, res) => {
   }
   res.status(200).end();
 };
-
-// exports.getImg = async (req, res) => {
-//   try {
-//     const { _id } = req.body;
-//     const user = await User.findOne({ _id });
-//     console.log(user);
-//     res.json(user.img);
-//   } catch (err) {
-//     console.error("Err message:", err.message);
-//     console.error("Err code", err);
-//   }
-//   res.status(200).end();
-// };
